corrected type

+++ lab/sinus_vs_square_hard/data.py
@@ -8,7 +8,7 @@ def create_data(nb_periods_train: int = 300, nb_periods_test: int = 10) -> Data:
     square = np.array([1, 1, 1, 1, -1, -1, -1, -1])
 
     X_train_base = np.empty((nb_periods_train, NB_POINTS_PER_PERIOD))
-    Y_train_base = np.random.binomial(1, 0.5, nb_periods_train).astype(np.float64)
+    Y_train_base = np.random.binomial(1, 0.5, nb_periods_train)
 
     for i in range(nb_periods_train):
         X_train_base[i, :] = sinus if Y_train_base[i] == 1 else square
@@ -16,8 +16,9 @@ def create_data(nb_periods_train: int = 300, nb_periods_test: int = 10) -> Data:
     X_train = X_train_base.reshape(-1, 1)
     Y_train = np.repeat(Y_train_base, NB_POINTS_PER_PERIOD).reshape(-1, 1)
 
+
     X_test_base = np.empty((nb_periods_test, NB_POINTS_PER_PERIOD))
-    Y_test_base = np.random.binomial(1, 0.5, (nb_periods_test, 1)).astype(np.float64)
+    Y_test_base = np.random.binomial(1, 0.5, (nb_periods_test, 1))
 
     for i in range(nb_periods_test):
         X_test_base[i, :] = sinus if Y_test_base[i] == 1 else square
